Update slideout.js
Add disabled elements.
Add "disabled" class of the parent element. And If you try to open a menu on this element (or a child), everything will fail.
I do not know how to post on githab right, I apologize. But I had no need to open menu on the slider.